util: Changed the type to u8 in associated with phys fields

+++ utility/register/option.hpp
@@ -13,11 +13,11 @@ namespace util {
         std::pair<uint8_t, uint8_t> bit_range;
         FixedP fixedp;
 
-        uint64_t hw_size() const {
+        uint8_t hw_size() const {
             return bit_range.second - bit_range.first + 1; // [4; 8] = 5bits 
         }
 
-        uint64_t log_size() const {
+        uint8_t log_size() const {
             return fixedp.int_bits + fixedp.fract_bits;
         }
